+++ src/ZBE_main/include/game/entities/GameBoard.h
@@ -23,7 +23,7 @@ namespace game {
 
 class GameBoard : public zbe::CollisionerEntity<GameReactor> {
 public:
-  GameBoard(double width, double height, uint64_t actuatorsList)
+  GameBoard(int64_t width, int64_t height, uint64_t actuatorsList)
     : c(new zbe::SimpleCollisioner<GameReactor>(std::make_shared<zbe::StaticLimiterAABB2D<GameReactor> >(zbe::AABB2D(zbe::AABB2D({0, 0}, {width, height} ))),
          std::make_shared<zbe::VoidReactObject<GameReactor> >(),
          actuatorsList)) {}
